Update jira components
