--- pkg/components/cloudcompute/otherprovider/capabilities.go
@@ -1,12 +1,0 @@
-package cloudcomputeotherprovider
-
-import (
-	v1 "github.com/openshift-eng/ci-test-mapping/pkg/api/types/v1"
-	"github.com/openshift-eng/ci-test-mapping/pkg/util"
-)
-
-func identifyCapabilities(test *v1.TestInfo) []string {
-	capabilities := util.DefaultCapabilities(test)
-
-	return capabilities
-}
