+++ pkg/registry/registry.go
@@ -16,16 +16,20 @@ import (
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/climanager"
 	cloudcomputebaremetalprovider "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/baremetalprovider"
 	cloudcomputecloudcontrollermanager "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/cloudcontrollermanager"
+	cloudcomputeclusterapiproviders "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/clusterapiproviders"
 	cloudcomputeclusterautoscaler "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/clusterautoscaler"
+	cloudcomputecontrolplanemachineset "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/controlplanemachineset"
 	cloudcomputeexternalprovider "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/externalprovider"
 	cloudcomputeibmprovider "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/ibmprovider"
 	cloudcomputekubevirtprovider "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/kubevirtprovider"
 	cloudcomputelibvirtprovider "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/libvirtprovider"
+	cloudcomputemachineapiproviders "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/machineapiproviders"
+	cloudcomputemachinecsrapprover "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/machinecsrapprover"
 	cloudcomputemachinehealthcheck "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/machinehealthcheck"
 	cloudcomputenutanixprovider "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/nutanixprovider"
 	cloudcomputeopenstackprovider "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/openstackprovider"
-	cloudcomputeotherprovider "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/otherprovider"
 	cloudcomputeovirtprovider "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/ovirtprovider"
+	cloudcomputeunknown "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcompute/unknown"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/cloudcredentialoperator"
 	cloudnativeeventscloudeventproxy "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudnativeevents/cloudeventproxy"
 	cloudnativeeventscloudnativeevents "github.com/openshift-eng/ci-test-mapping/pkg/components/cloudnativeevents/cloudnativeevents"
@@ -35,6 +39,7 @@ import (
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/clusterversionoperator"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/cnfcerttnf"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/complianceoperator"
+	"github.com/openshift-eng/ci-test-mapping/pkg/components/confidentialcomputeattestation"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/configoperator"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/consolemetal3plugin"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/consolestorageplugin"
@@ -123,6 +128,7 @@ import (
 	nodecpumanager "github.com/openshift-eng/ci-test-mapping/pkg/components/node/cpumanager"
 	nodecrio "github.com/openshift-eng/ci-test-mapping/pkg/components/node/crio"
 	nodedevicemanager "github.com/openshift-eng/ci-test-mapping/pkg/components/node/devicemanager"
+	nodeinstasliceoperator "github.com/openshift-eng/ci-test-mapping/pkg/components/node/instasliceoperator"
 	nodekubelet "github.com/openshift-eng/ci-test-mapping/pkg/components/node/kubelet"
 	nodememorymanager "github.com/openshift-eng/ci-test-mapping/pkg/components/node/memorymanager"
 	nodenodeproblemdetector "github.com/openshift-eng/ci-test-mapping/pkg/components/node/nodeproblemdetector"
@@ -139,6 +145,7 @@ import (
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/oauthproxy"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/observabilityui"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/oc"
+	occlustercompare "github.com/openshift-eng/ci-test-mapping/pkg/components/oc/clustercompare"
 	ococmirror "github.com/openshift-eng/ci-test-mapping/pkg/components/oc/ocmirror"
 	ocupdate "github.com/openshift-eng/ci-test-mapping/pkg/components/oc/update"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/occompliance"
@@ -190,7 +197,6 @@ import (
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/testframework"
 	testframeworkopenstack "github.com/openshift-eng/ci-test-mapping/pkg/components/testframework/openstack"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/testinfrastructure"
-	"github.com/openshift-eng/ci-test-mapping/pkg/components/topolvm"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/unknown"
 	"github.com/openshift-eng/ci-test-mapping/pkg/components/windowscontainers"
 )
@@ -202,50 +208,56 @@ type Registry struct {
 func NewComponentRegistry() *Registry {
 	var r Registry
 
-	r.Register("apiserver-auth", &apiserverauth.ApiserverAuthComponent)
 	r.Register("AWS Load Balancer Operator", &awsloadbalanceroperator.AWSLoadBalancerOperatorComponent)
-	r.Register("Bare Metal Hardware Provisioning", &baremetalhardwareprovisioning.BareMetalHardwareProvisioningComponent)
+	r.Register("BMER Events", &bmerevents.BMEREventsComponent)
+	r.Register("Bare Metal Hardware Provisioning / OS Image Provider", &baremetalhardwareprovisioningosimageprovider.OSImageProviderComponent)
 	r.Register("Bare Metal Hardware Provisioning / baremetal-operator", &baremetalhardwareprovisioningbaremetaloperator.BaremetalOperatorComponent)
 	r.Register("Bare Metal Hardware Provisioning / cluster-api-provider", &baremetalhardwareprovisioningclusterapiprovider.ClusterAPIProviderComponent)
 	r.Register("Bare Metal Hardware Provisioning / cluster-baremetal-operator", &baremetalhardwareprovisioningclusterbaremetaloperator.ClusterBaremetalOperatorComponent)
 	r.Register("Bare Metal Hardware Provisioning / ironic", &baremetalhardwareprovisioningironic.IronicComponent)
-	r.Register("Bare Metal Hardware Provisioning / OS Image Provider", &baremetalhardwareprovisioningosimageprovider.OSImageProviderComponent)
+	r.Register("Bare Metal Hardware Provisioning", &baremetalhardwareprovisioning.BareMetalHardwareProvisioningComponent)
 	r.Register("Build", &build.BuildComponent)
-	r.Register("cert-manager", &certmanager.CertManagerComponent)
+	r.Register("CNF-Cert-TNF", &cnfcerttnf.CNFCertTNFComponent)
 	r.Register("Cloud Compute / BareMetal Provider", &cloudcomputebaremetalprovider.BareMetalProviderComponent)
 	r.Register("Cloud Compute / Cloud Controller Manager", &cloudcomputecloudcontrollermanager.CloudControllerManagerComponent)
+	r.Register("Cloud Compute / Cluster API Providers", &cloudcomputeclusterapiproviders.ClusterAPIProvidersComponent)
 	r.Register("Cloud Compute / Cluster Autoscaler", &cloudcomputeclusterautoscaler.ClusterAutoscalerComponent)
+	r.Register("Cloud Compute / ControlPlaneMachineSet", &cloudcomputecontrolplanemachineset.ControlPlaneMachineSetComponent)
+	r.Register("Cloud Compute / External Provider", &cloudcomputeexternalprovider.ExternalProviderComponent)
 	r.Register("Cloud Compute / IBM Provider", &cloudcomputeibmprovider.IBMProviderComponent)
 	r.Register("Cloud Compute / KubeVirt Provider", &cloudcomputekubevirtprovider.KubeVirtProviderComponent)
+	r.Register("Cloud Compute / Libvirt Provider", &cloudcomputelibvirtprovider.LibvirtProviderComponent)
+	r.Register("Cloud Compute / Machine API Providers", &cloudcomputemachineapiproviders.MachineAPIProvidersComponent)
+	r.Register("Cloud Compute / Machine CSR Approver", &cloudcomputemachinecsrapprover.MachineCSRApproverComponent)
 	r.Register("Cloud Compute / MachineHealthCheck", &cloudcomputemachinehealthcheck.MachineHealthCheckComponent)
 	r.Register("Cloud Compute / Nutanix Provider", &cloudcomputenutanixprovider.NutanixProviderComponent)
 	r.Register("Cloud Compute / OpenStack Provider", &cloudcomputeopenstackprovider.OpenStackProviderComponent)
-	r.Register("Cloud Compute / Other Provider", &cloudcomputeotherprovider.OtherProviderComponent)
+	r.Register("Cloud Compute / Unknown", &cloudcomputeunknown.UnknownComponent)
 	r.Register("Cloud Compute / oVirt Provider", &cloudcomputeovirtprovider.OVirtProviderComponent)
 	r.Register("Cloud Credential Operator", &cloudcredentialoperator.CloudCredentialOperatorComponent)
 	r.Register("Cloud Native Events / Cloud Event Proxy", &cloudnativeeventscloudeventproxy.CloudEventProxyComponent)
 	r.Register("Cloud Native Events / Cloud Native Events", &cloudnativeeventscloudnativeevents.CloudNativeEventsComponent)
 	r.Register("Cloud Native Events / Hardware Event Proxy", &cloudnativeeventshardwareeventproxy.HardwareEventProxyComponent)
 	r.Register("Cluster Loader", &clusterloader.ClusterLoaderComponent)
+	r.Register("Cluster Resource Override Admission Operator", &clusterresourceoverrideadmissionoperator.ClusterResourceOverrideAdmissionOperatorComponent)
 	r.Register("Cluster Version Operator", &clusterversionoperator.ClusterVersionOperatorComponent)
-	r.Register("CNF-Cert-TNF", &cnfcerttnf.CNFCertTNFComponent)
 	r.Register("Compliance Operator", &complianceoperator.ComplianceOperatorComponent)
-	r.Register("config-operator", &configoperator.ConfigOperatorComponent)
 	r.Register("Console Metal3 Plugin", &consolemetal3plugin.ConsoleMetal3PluginComponent)
 	r.Register("Console Storage Plugin", &consolestorageplugin.ConsoleStoragePluginComponent)
 	r.Register("Containers", &containers.ContainersComponent)
-	r.Register("crc", &crc.CrcComponent)
-	r.Register("descheduler", &descheduler.DeschedulerComponent)
 	r.Register("Dev Console", &devconsole.DevConsoleComponent)
 	r.Register("Driver Toolkit", &drivertoolkit.DriverToolkitComponent)
 	r.Register("Etcd", &etcd.EtcdComponent)
 	r.Register("ExternalDNS Operator", &externaldnsoperator.ExternalDNSOperatorComponent)
 	r.Register("File Integrity Operator", &fileintegrityoperator.FileIntegrityOperatorComponent)
+	r.Register("GitOps ZTP", &gitopsztp.GitOpsZTPComponent)
 	r.Register("Hawkular", &hawkular.HawkularComponent)
 	r.Register("Helm", &helm.HelmComponent)
 	r.Register("Hive", &hive.HiveComponent)
+	r.Register("HyperShift / Agent", &hypershiftagent.AgentComponent)
+	r.Register("HyperShift / OCP Virtualization", &hypershiftocpvirtualization.OCPVirtualizationComponent)
 	r.Register("HyperShift", &hypershift.HyperShiftComponent)
-	r.Register("ibm-roks-toolkit", &ibmrokstoolkit.IbmRoksToolkitComponent)
+	r.Register("ISV Operators", &isvoperators.ISVOperatorsComponent)
 	r.Register("Image Registry", &imageregistry.ImageRegistryComponent)
 	r.Register("ImageStreams", &imagestreams.ImageStreamsComponent)
 	r.Register("Insights Operator", &insightsoperator.InsightsOperatorComponent)
@@ -257,54 +269,60 @@ func NewComponentRegistry() *Registry {
 	r.Register("Installer / OpenShift on Bare Metal IPI", &installeropenshiftonbaremetalipi.OpenShiftOnBareMetalIPIComponent)
 	r.Register("Installer / OpenShift on OpenStack", &installeropenshiftonopenstack.OpenShiftOnOpenStackComponent)
 	r.Register("Installer / OpenShift on RHV", &installeropenshiftonrhv.OpenShiftOnRHVComponent)
-	r.Register("Installer / openshift-ansible", &installeropenshiftansible.OpenshiftAnsibleComponent)
-	r.Register("Installer / openshift-installer", &installeropenshiftinstaller.OpenshiftInstallerComponent)
 	r.Register("Installer / PowerVS", &installerpowervs.PowerVSComponent)
 	r.Register("Installer / Single Node OpenShift", &installersinglenodeopenshift.SingleNodeOpenShiftComponent)
-	r.Register("ISV Operators", &isvoperators.ISVOperatorsComponent)
+	r.Register("Installer / openshift-ansible", &installeropenshiftansible.OpenshiftAnsibleComponent)
+	r.Register("Installer / openshift-installer", &installeropenshiftinstaller.OpenshiftInstallerComponent)
 	r.Register("Jenkins", &jenkins.JenkinsComponent)
-	r.Register("kmm", &kmm.KmmComponent)
-	r.Register("kube-apiserver", &kubeapiserver.KubeApiserverComponent)
-	r.Register("kube-controller-manager", &kubecontrollermanager.KubeControllerManagerComponent)
-	r.Register("kube-scheduler", &kubescheduler.KubeSchedulerComponent)
-	r.Register("kube-storage-version-migrator", &kubestorageversionmigrator.KubeStorageVersionMigratorComponent)
+	r.Register("LCA operator", &lcaoperator.LCAOperatorComponent)
 	r.Register("Logging", &logging.LoggingComponent)
-	r.Register("Machine Config Operator", &machineconfigoperator.MachineConfigOperatorComponent)
+	r.Register("Logical Volume Manager Storage", &logicalvolumemanagerstorage.LogicalVolumeManagerStorageComponent)
+	r.Register("Low latency validation tooling", &lowlatencyvalidationtooling.LowLatencyValidationToolingComponent)
 	r.Register("Machine Config Operator / platform-baremetal", &machineconfigoperatorplatformbaremetal.PlatformBaremetalComponent)
 	r.Register("Machine Config Operator / platform-none", &machineconfigoperatorplatformnone.PlatformNoneComponent)
 	r.Register("Machine Config Operator / platform-openstack", &machineconfigoperatorplatformopenstack.PlatformOpenstackComponent)
 	r.Register("Machine Config Operator / platform-ovirt-rhv", &machineconfigoperatorplatformovirtrhv.PlatformOvirtRhvComponent)
 	r.Register("Machine Config Operator / platform-vsphere", &machineconfigoperatorplatformvsphere.PlatformVsphereComponent)
+	r.Register("Machine Config Operator", &machineconfigoperator.MachineConfigOperatorComponent)
 	r.Register("Management Console", &managementconsole.ManagementConsoleComponent)
 	r.Register("Metering Operator", &meteringoperator.MeteringOperatorComponent)
-	r.Register("MicroShift", &microshift.MicroShiftComponent)
 	r.Register("MicroShift / Networking", &microshiftnetworking.NetworkingComponent)
 	r.Register("MicroShift / Storage", &microshiftstorage.StorageComponent)
-	r.Register("Monitoring", &monitoring.MonitoringComponent)
+	r.Register("MicroShift", &microshift.MicroShiftComponent)
 	r.Register("Monitoring / Grafana", &monitoringgrafana.GrafanaComponent)
-	r.Register("Multi-Arch", &multiarch.MultiArchComponent)
+	r.Register("Monitoring", &monitoring.MonitoringComponent)
 	r.Register("Multi-Arch / ARM", &multiarcharm.ARMComponent)
 	r.Register("Multi-Arch / IBM P and Z", &multiarchibmpandz.IBMPAndZComponent)
+	r.Register("Multi-Arch", &multiarch.MultiArchComponent)
+	r.Register("Multiarch Tuning Operator", &multiarchtuningoperator.MultiarchTuningOperatorComponent)
+	r.Register("NVIDIA", &nvidia.NVIDIAComponent)
+	r.Register("Networking / DNS", &networkingdns.DNSComponent)
+	r.Register("Networking / DPU", &networkingdpu.DPUComponent)
+	r.Register("Networking / Metal LB", &networkingmetallb.MetalLBComponent)
+	r.Register("Networking / NetObs", &networkingnetobs.NetObsComponent)
+	r.Register("Networking / On-Prem DNS", &networkingonpremdns.OnPremDNSComponent)
+	r.Register("Networking / On-Prem Host Networking", &networkingonpremhostnetworking.OnPremHostNetworkingComponent)
+	r.Register("Networking / On-Prem Load Balancer", &networkingonpremloadbalancer.OnPremLoadBalancerComponent)
+	r.Register("Networking / SR-IOV", &networkingsriov.SRIOVComponent)
 	r.Register("Networking / cloud-network-config-controller", &networkingcloudnetworkconfigcontroller.CloudNetworkConfigControllerComponent)
 	r.Register("Networking / cluster-network-operator", &networkingclusternetworkoperator.ClusterNetworkOperatorComponent)
-	r.Register("Networking / DNS", &networkingdns.DNSComponent)
 	r.Register("Networking / ingress-node-firewall", &networkingingressnodefirewall.IngressNodeFirewallComponent)
 	r.Register("Networking / kubernetes-nmstate", &networkingkubernetesnmstate.KubernetesNmstateComponent)
 	r.Register("Networking / kubernetes-nmstate-operator", &networkingkubernetesnmstateoperator.KubernetesNmstateOperatorComponent)
 	r.Register("Networking / kuryr", &networkingkuryr.KuryrComponent)
-	r.Register("Networking / Metal LB", &networkingmetallb.MetalLBComponent)
 	r.Register("Networking / multus", &networkingmultus.MultusComponent)
-	r.Register("Networking / NetObs", &networkingnetobs.NetObsComponent)
+	r.Register("Networking / network-tools", &networkingnetworktools.NetworkToolsComponent)
+	r.Register("Networking / networking-console-plugin", &networkingnetworkingconsoleplugin.NetworkingConsolePluginComponent)
 	r.Register("Networking / nmstate-console-plugin", &networkingnmstateconsoleplugin.NmstateConsolePluginComponent)
 	r.Register("Networking / openshift-sdn", &networkingopenshiftsdn.OpenshiftSdnComponent)
 	r.Register("Networking / ovn-kubernetes", &networkingovnkubernetes.OvnKubernetesComponent)
 	r.Register("Networking / ptp", &networkingptp.PtpComponent)
 	r.Register("Networking / router", &networkingrouter.RouterComponent)
 	r.Register("Networking / runtime-cfg", &networkingruntimecfg.RuntimeCfgComponent)
-	r.Register("Networking / SR-IOV", &networkingsriov.SRIOVComponent)
 	r.Register("Node / CPU manager", &nodecpumanager.CPUManagerComponent)
 	r.Register("Node / CRI-O", &nodecrio.CRIOComponent)
 	r.Register("Node / Device Manager", &nodedevicemanager.DeviceManagerComponent)
+	r.Register("Node / Instaslice-operator", &nodeinstasliceoperator.InstasliceOperatorComponent)
 	r.Register("Node / Kubelet", &nodekubelet.KubeletComponent)
 	r.Register("Node / Memory manager", &nodememorymanager.MemoryManagerComponent)
 	r.Register("Node / Node Problem Detector", &nodenodeproblemdetector.NodeProblemDetectorComponent)
@@ -315,84 +333,78 @@ func NewComponentRegistry() *Registry {
 	r.Register("Node Maintenance Operator", &nodemaintenanceoperator.NodeMaintenanceOperatorComponent)
 	r.Register("Node Tuning Operator", &nodetuningoperator.NodeTuningOperatorComponent)
 	r.Register("Node-observability-operator", &nodeobservabilityoperator.NodeObservabilityOperatorComponent)
-	r.Register("none", &none.NoneComponent)
-	r.Register("oauth-apiserver", &oauthapiserver.OauthApiserverComponent)
-	r.Register("oauth-proxy", &oauthproxy.OauthProxyComponent)
-	r.Register("Observability UI", &observabilityui.ObservabilityUIComponent)
-	r.Register("oc", &oc.OcComponent)
-	r.Register("oc / oc-mirror", &ococmirror.OcMirrorComponent)
-	r.Register("oc-compliance", &occompliance.OcComplianceComponent)
-	r.Register("OLM", &olm.OLMComponent)
 	r.Register("OLM / OperatorHub", &olmoperatorhub.OperatorHubComponent)
 	r.Register("OLM / Registry", &olmregistry.RegistryComponent)
-	r.Register("OpenShift Update Service / operand", &openshiftupdateserviceoperand.OperandComponent)
-	r.Register("OpenShift Update Service / operator", &openshiftupdateserviceoperator.OperatorComponent)
-	r.Register("openshift-apiserver", &openshiftapiserver.OpenshiftApiserverComponent)
-	r.Register("openshift-controller-manager / apps", &openshiftcontrollermanagerapps.AppsComponent)
-	r.Register("openshift-controller-manager / build", &openshiftcontrollermanagerbuild.BuildComponent)
-	r.Register("openshift-controller-manager / controller-manager", &openshiftcontrollermanagercontrollermanager.ControllerManagerComponent)
-	r.Register("Operator SDK", &operatorsdk.OperatorSDKComponent)
-	r.Register("Performance Addon Operator", &performanceaddonoperator.PerformanceAddonOperatorComponent)
-	r.Register("Poison Pill Operator", &poisonpilloperator.PoisonPillOperatorComponent)
-	r.Register("Registry Console", &registryconsole.RegistryConsoleComponent)
-	r.Register("Release", &release.ReleaseComponent)
-	r.Register("RHCOS", &rhcos.RHCOSComponent)
-	r.Register("RHMI Monitoring", &rhmimonitoring.RHMIMonitoringComponent)
-	r.Register("route-controller-manager", &routecontrollermanager.RouteControllerManagerComponent)
-	r.Register("run-once-duration-override", &runoncedurationoverride.RunOnceDurationOverrideComponent)
-	r.Register("Samples Operator", &samplesoperator.SamplesOperatorComponent)
-	r.Register("sandboxed-containers", &sandboxedcontainers.SandboxedContainersComponent)
-	r.Register("secondary-scheduler-operator", &secondaryscheduleroperator.SecondarySchedulerOperatorComponent)
-	r.Register("Security", &security.SecurityComponent)
-	r.Register("Security Profiles Operator", &securityprofilesoperator.SecurityProfilesOperatorComponent)
-	r.Register("Service Binding", &servicebinding.ServiceBindingComponent)
-	r.Register("Service Broker", &servicebroker.ServiceBrokerComponent)
-	r.Register("Service Catalog", &servicecatalog.ServiceCatalogComponent)
-	r.Register("service-ca", &serviceca.ServiceCaComponent)
-	r.Register("Special Resource Operator", &specialresourceoperator.SpecialResourceOperatorComponent)
-	r.Register("Storage", &storage.StorageComponent)
-	r.Register("Storage / Kubernetes", &storagekubernetes.KubernetesComponent)
-	r.Register("Storage / Kubernetes External Components", &storagekubernetesexternalcomponents.KubernetesExternalComponentsComponent)
-	r.Register("Storage / Local Storage Operator", &storagelocalstorageoperator.LocalStorageOperatorComponent)
-	r.Register("Storage / OpenStack CSI Drivers", &storageopenstackcsidrivers.OpenStackCSIDriversComponent)
-	r.Register("Storage / Operators", &storageoperators.OperatorsComponent)
-	r.Register("Storage / oVirt CSI Driver", &storageovirtcsidriver.OVirtCSIDriverComponent)
-	r.Register("Storage / Shared Resource CSI Driver", &storagesharedresourcecsidriver.SharedResourceCSIDriverComponent)
-	r.Register("Telemeter", &telemeter.TelemeterComponent)
-	r.Register("Templates", &templates.TemplatesComponent)
-	r.Register("Test Framework", &testframework.TestFrameworkComponent)
-	r.Register("Test Framework / OpenStack", &testframeworkopenstack.OpenStackComponent)
-	r.Register("Test Infrastructure", &testinfrastructure.TestInfrastructureComponent)
-	r.Register("topolvm", &topolvm.TopolvmComponent)
-	r.Register("Unknown", &unknown.UnknownComponent)
-	r.Register("Windows Containers", &windowscontainers.WindowsContainersComponent)
-	r.Register("Cloud Compute / External Provider", &cloudcomputeexternalprovider.ExternalProviderComponent)
-	r.Register("Cluster Resource Override Admission Operator", &clusterresourceoverrideadmissionoperator.ClusterResourceOverrideAdmissionOperatorComponent)
-	r.Register("Networking / network-tools", &networkingnetworktools.NetworkToolsComponent)
-	r.Register("Networking / On-Prem DNS", &networkingonpremdns.OnPremDNSComponent)
-	r.Register("Networking / On-Prem Host Networking", &networkingonpremhostnetworking.OnPremHostNetworkingComponent)
-	r.Register("Networking / On-Prem Load Balancer", &networkingonpremloadbalancer.OnPremLoadBalancerComponent)
-	r.Register("NVIDIA", &nvidia.NVIDIAComponent)
-	r.Register("Pod Autoscaler", &podautoscaler.PodAutoscalerComponent)
-	r.Register("BMER Events", &bmerevents.BMEREventsComponent)
-	r.Register("cli-manager", &climanager.CliManagerComponent)
-	r.Register("Cloud Compute / Libvirt Provider", &cloudcomputelibvirtprovider.LibvirtProviderComponent)
-	r.Register("GitOps ZTP", &gitopsztp.GitOpsZTPComponent)
-	r.Register("HyperShift / Agent", &hypershiftagent.AgentComponent)
-	r.Register("HyperShift / OCP Virtualization", &hypershiftocpvirtualization.OCPVirtualizationComponent)
-	r.Register("LCA operator", &lcaoperator.LCAOperatorComponent)
-	r.Register("Logical Volume Manager Storage", &logicalvolumemanagerstorage.LogicalVolumeManagerStorageComponent)
-	r.Register("Low latency validation tooling", &lowlatencyvalidationtooling.LowLatencyValidationToolingComponent)
-	r.Register("Multiarch Tuning Operator", &multiarchtuningoperator.MultiarchTuningOperatorComponent)
-	r.Register("Networking / DPU", &networkingdpu.DPUComponent)
-	r.Register("Networking / networking-console-plugin", &networkingnetworkingconsoleplugin.NetworkingConsolePluginComponent)
-	r.Register("oc / update", &ocupdate.UpdateComponent)
+	r.Register("OLM", &olm.OLMComponent)
 	r.Register("OPCT / CLI", &opctcli.CLIComponent)
 	r.Register("OPCT / Other", &opctother.OtherComponent)
 	r.Register("OPCT / Results / General", &opctresultsgeneral.GeneralComponent)
 	r.Register("OPCT / Results / VCSP", &opctresultsvcsp.VCSPComponent)
+	r.Register("Observability UI", &observabilityui.ObservabilityUIComponent)
+	r.Register("OpenShift Update Service / operand", &openshiftupdateserviceoperand.OperandComponent)
+	r.Register("OpenShift Update Service / operator", &openshiftupdateserviceoperator.OperatorComponent)
+	r.Register("Operator SDK", &operatorsdk.OperatorSDKComponent)
+	r.Register("Performance Addon Operator", &performanceaddonoperator.PerformanceAddonOperatorComponent)
+	r.Register("Pod Autoscaler", &podautoscaler.PodAutoscalerComponent)
+	r.Register("Poison Pill Operator", &poisonpilloperator.PoisonPillOperatorComponent)
+	r.Register("RHCOS", &rhcos.RHCOSComponent)
+	r.Register("RHMI Monitoring", &rhmimonitoring.RHMIMonitoringComponent)
+	r.Register("Registry Console", &registryconsole.RegistryConsoleComponent)
+	r.Register("Release", &release.ReleaseComponent)
+	r.Register("Samples Operator", &samplesoperator.SamplesOperatorComponent)
+	r.Register("Security Profiles Operator", &securityprofilesoperator.SecurityProfilesOperatorComponent)
+	r.Register("Security", &security.SecurityComponent)
+	r.Register("Service Binding", &servicebinding.ServiceBindingComponent)
+	r.Register("Service Broker", &servicebroker.ServiceBrokerComponent)
+	r.Register("Service Catalog", &servicecatalog.ServiceCatalogComponent)
+	r.Register("Special Resource Operator", &specialresourceoperator.SpecialResourceOperatorComponent)
+	r.Register("Storage / Kubernetes External Components", &storagekubernetesexternalcomponents.KubernetesExternalComponentsComponent)
+	r.Register("Storage / Kubernetes", &storagekubernetes.KubernetesComponent)
+	r.Register("Storage / Local Storage Operator", &storagelocalstorageoperator.LocalStorageOperatorComponent)
+	r.Register("Storage / OpenStack CSI Drivers", &storageopenstackcsidrivers.OpenStackCSIDriversComponent)
+	r.Register("Storage / Operators", &storageoperators.OperatorsComponent)
+	r.Register("Storage / Shared Resource CSI Driver", &storagesharedresourcecsidriver.SharedResourceCSIDriverComponent)
+	r.Register("Storage / oVirt CSI Driver", &storageovirtcsidriver.OVirtCSIDriverComponent)
+	r.Register("Storage", &storage.StorageComponent)
 	r.Register("TALM Operator", &talmoperator.TALMOperatorComponent)
 	r.Register("Telco Performance", &telcoperformance.TelcoPerformanceComponent)
+	r.Register("Telemeter", &telemeter.TelemeterComponent)
+	r.Register("Templates", &templates.TemplatesComponent)
+	r.Register("Test Framework / OpenStack", &testframeworkopenstack.OpenStackComponent)
+	r.Register("Test Framework", &testframework.TestFrameworkComponent)
+	r.Register("Test Infrastructure", &testinfrastructure.TestInfrastructureComponent)
+	r.Register("Unknown", &unknown.UnknownComponent)
+	r.Register("Windows Containers", &windowscontainers.WindowsContainersComponent)
+	r.Register("apiserver-auth", &apiserverauth.ApiserverAuthComponent)
+	r.Register("cert-manager", &certmanager.CertManagerComponent)
+	r.Register("cli-manager", &climanager.CliManagerComponent)
+	r.Register("confidential-compute-attestation", &confidentialcomputeattestation.ConfidentialComputeAttestationComponent)
+	r.Register("config-operator", &configoperator.ConfigOperatorComponent)
+	r.Register("crc", &crc.CrcComponent)
+	r.Register("descheduler", &descheduler.DeschedulerComponent)
+	r.Register("ibm-roks-toolkit", &ibmrokstoolkit.IbmRoksToolkitComponent)
+	r.Register("kmm", &kmm.KmmComponent)
+	r.Register("kube-apiserver", &kubeapiserver.KubeApiserverComponent)
+	r.Register("kube-controller-manager", &kubecontrollermanager.KubeControllerManagerComponent)
+	r.Register("kube-scheduler", &kubescheduler.KubeSchedulerComponent)
+	r.Register("kube-storage-version-migrator", &kubestorageversionmigrator.KubeStorageVersionMigratorComponent)
+	r.Register("none", &none.NoneComponent)
+	r.Register("oauth-apiserver", &oauthapiserver.OauthApiserverComponent)
+	r.Register("oauth-proxy", &oauthproxy.OauthProxyComponent)
+	r.Register("oc / cluster-compare", &occlustercompare.ClusterCompareComponent)
+	r.Register("oc / oc-mirror", &ococmirror.OcMirrorComponent)
+	r.Register("oc / update", &ocupdate.UpdateComponent)
+	r.Register("oc", &oc.OcComponent)
+	r.Register("oc-compliance", &occompliance.OcComplianceComponent)
+	r.Register("openshift-apiserver", &openshiftapiserver.OpenshiftApiserverComponent)
+	r.Register("openshift-controller-manager / apps", &openshiftcontrollermanagerapps.AppsComponent)
+	r.Register("openshift-controller-manager / build", &openshiftcontrollermanagerbuild.BuildComponent)
+	r.Register("openshift-controller-manager / controller-manager", &openshiftcontrollermanagercontrollermanager.ControllerManagerComponent)
+	r.Register("route-controller-manager", &routecontrollermanager.RouteControllerManagerComponent)
+	r.Register("run-once-duration-override", &runoncedurationoverride.RunOnceDurationOverrideComponent)
+	r.Register("sandboxed-containers", &sandboxedcontainers.SandboxedContainersComponent)
+	r.Register("secondary-scheduler-operator", &secondaryscheduleroperator.SecondarySchedulerOperatorComponent)
+	r.Register("service-ca", &serviceca.ServiceCaComponent)
 	// New components go here
 
 	return &r
